frontend/requirements.txt
llama-index-core>=0.10.0
llama-index-embeddings-huggingface>=0.1.3
chromadb>=0.4.24
PyMuPDF>=1.23.0
tqdm>=4.66.0
python-dotenv>=1.0.0
streamlit>=1.31.0
requests>=2.31.0

